Fixed MSVC build errors and updated projects.

+++ libserpent/util.cpp
@@ -205,7 +205,7 @@ std::string get_file_contents(std::string filename)
   {
     std::string contents;
     in.seekg(0, std::ios::end);
-    contents.resize(in.tellg());
+    contents.resize((size_t)in.tellg());
     in.seekg(0, std::ios::beg);
     in.read(&contents[0], contents.size());
     in.close();
